server/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
starlette==0.37.2
httpx==0.27.0
pydantic==2.7.1
google-auth # Required for Google ID token fetching
requests 
